+++ table/retry.go
@@ -113,38 +113,54 @@ func Retry(ctx context.Context, s SessionProvider, op Operation) error {
 
 // Do calls op.Do until it return nil or not retriable error.
 func (r Retryer) sessionGet(ctx context.Context) (_ *Session, err error) {
-	result := make(chan struct {
+	if err = ctx.Err(); err != nil {
+		return nil, err // simple call saves a goroutine
+	}
+	type result struct {
 		s *Session
 		e error
-	})
+	}
+	res := make(chan result)
 	go func() {
-		defer close(result)
+		defer close(res)
 		s, e := r.SessionProvider.Get(ctx)
-		result <- struct {
-			s *Session
-			e error
-		}{s: s, e: e}
+		if s == nil && e == nil {
+			panic("only one of pair <session, error> must be not nil")
+		}
+		select {
+		case res <- result{s: s, e: e}:
+		default:
+			if s != nil {
+				_ = r.SessionProvider.Put(ctx, s)
+			}
+		}
 	}()
 	select {
 	case <-ctx.Done():
 		return nil, ctx.Err()
-	case result := <-result:
-		return result.s, result.e
+	case r, ok := <-res:
+		if !ok {
+			return nil, context.DeadlineExceeded
+		}
+		return r.s, r.e
 	}
 }
 
 // Do calls op.Do until it return nil or not retriable error.
-func (r Retryer) opDo(ctx context.Context, op Operation, s *Session) error {
-	err := make(chan error)
+func (r Retryer) opDo(ctx context.Context, op Operation, s *Session) (err error) {
+	if err = ctx.Err(); err != nil {
+		return err // simple call saves a goroutine
+	}
+	errs := make(chan error, 1)
 	go func() {
-		defer close(err)
-		err <- op.Do(ctx, s)
+		defer close(errs)
+		errs <- op.Do(ctx, s)
 	}()
 	select {
 	case <-ctx.Done():
 		return ctx.Err()
-	case e := <-err:
-		return e
+	case err = <-errs:
+		return err
 	}
 }